requirements.txt
bcrypt==4.0.1
customtkinter==5.1.2
darkdetect==0.8.0
mysql-connector-python==8.0.32
Pillow==9.4.0
protobuf==3.20.3
tk==0.1.0
ttkthemes==3.2.2
